refactor: Improve architecture with dependency injection and type safety
- Refactor Input system from static to instance-based with dependency injection
- Remove const_cast from Registry::sort() by adding non-const entities() getter
- Improve pointer-to-integer conversion in SelectableId using uintptr_t
- Implement Shader::createFromFile for loading shaders from disk

+++ src/esengine/editor/core/Selection.hpp
@@ -57,7 +57,8 @@ struct SelectableId {
         if constexpr (std::is_integral_v<T>) {
             id.rawId = static_cast<u64>(value);
         } else if constexpr (std::is_pointer_v<T>) {
-            id.rawId = reinterpret_cast<u64>(value);
+            // Use uintptr_t as intermediate type for safer pointer-to-integer conversion
+            id.rawId = static_cast<u64>(reinterpret_cast<uintptr_t>(value));
         } else {
             id.rawId = std::hash<T>{}(value);
         }
